fix diff edit save preconditions
